inference/requirements.txt
onnxruntime
numpy
opencv-python
Pillow
inference-sdk
supervision
